+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.8.0 (tbd)
 
 - Fixed pinning of dependencies in `piral-configs` (#733)
+- Improved error overlay with source maps and auto-close on reload
 - Updated `kras` to latest `v1.0` release
 - Updated `dets` to latest `v1.0` release
 - Updated dependencies